fixed backend/lib/postgres/ dependencies

+++ backend/lib/postgres/subscription_plans/SubscriptionPlanRepo.scala
@@ -6,7 +6,7 @@ import scala.concurrent.Future
 class SubscriptionPlanRepository(db: Database) {
     private val subscription_plans = TableQuery[SubscriptionPlanTable]
 
-    def findById(id: Long): Future[Option[SubscriptionPlan]] = 
+    def findById(id: Int): Future[Option[SubscriptionPlan]] = 
         db.run(subscription_plans.filter(_.spl_id === id).result.headOption)
 
     def findAll(): Future[Seq[SubscriptionPlan]] =
